requirements.txt
requests
boto3
scrapy
scrapy-rotating-proxies
python-dotenv
brotli